api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.1
fastapi-slim==0.111.0
h11==0.14.0
idna==3.7
pip==22.0.2
pydantic==2.7.4
pydantic-core==2.18.4
setuptools==59.6.0
sniffio==1.3.1
starlette==0.37.2
typing-extensions==4.12.2
uv==0.2.11
uvicorn==0.30.1
